Move MoveStats into a separate module
